refactor&test: add Message class and tests for every module
* refactor: add message class

Created message class to describe the message sent or received

* refactor: add pyrogram Message class

Using pyrogram message class to create custom message class.
this class should be the building block of this bot, every single message that the bot sends or receives must pass this class.

* test: add test module

Created tests for Lectures module and groups module (basically every module)
Hoping this will fasten the debugging of this project in the future.